requirements.txt
Flask==0.12.2
Flask-MongoAlchemy==0.7.2
cognitive-face==1.3.1
Werkzeug==0.12.2
pymongo==2.8.1
google-api-python-client
gunicorn==19.7.1
azure-storage-file==0.37.0
google-cloud-storage==1.3.2
